+++ src/Services/LikeableService.php
@@ -294,6 +294,7 @@ class LikeableService implements LikeableServiceContract
         $model = $query->getModel();
 
         return $query
+            ->select($model->getTable() . '.*', 'like_counter.count')
             ->leftJoin('like_counter', function ($join) use ($model) {
                 $join->on('like_counter.likeable_id', '=', "{$model->getTable()}.{$model->getKeyName()}");
                 $join->on('like_counter.likeable_type', '=', "{$model->getMorphClass()}");